fix(ui): version history now inside dialog next to preview
