--- ui/src/features/Editor/components/RightPanel/components/index.ts
@@ -1,1 +1,0 @@
-export * from "./VersionHistory";
