+++ ui/src/features/Editor/components/RightPanel/index.tsx
@@ -3,10 +3,7 @@ import { memo } from "react";
 import { Doc } from "yjs";
 
 import { IconButton } from "@flow/components";
-import { useT } from "@flow/lib/i18n";
 import type { Project } from "@flow/types";
-
-import { VersionHistoryList } from "./components";
 
 type Props = {
   contentType?: "version-history";
@@ -15,13 +12,7 @@ type Props = {
   yDoc: Doc | null;
 };
 
-const RightPanel: React.FC<Props> = ({
-  contentType,
-  onClose,
-  project,
-  yDoc,
-}) => {
-  const t = useT();
+const RightPanel: React.FC<Props> = ({ contentType, onClose }) => {
   return (
     <div
       id="right-panel"
@@ -38,13 +29,7 @@ const RightPanel: React.FC<Props> = ({
           icon={<X className="size-[18px]" weight="thin" />}
           onClick={onClose}
         />
-        <div className="grow p-2 text-center">
-          <p className="text-lg dark:font-thin">{t("Version History")}</p>
-        </div>
       </div>
-      {contentType === "version-history" && (
-        <VersionHistoryList project={project} yDoc={yDoc} />
-      )}
     </div>
   );
 };
