+++ ui/src/features/Editor/components/TopBar/components/ActionBar/components/Version/Version.tsx
@@ -20,7 +20,7 @@ const Version: React.FC<VersionProps> = ({
   return (
     <div>
       <div
-        className={`flex cursor-pointer select-none justify-between gap-2 rounded px-4 py-2 ${isSelected ? "relative before:absolute before:inset-y-0 before:left-0 before:right-1 before:-z-10 before:rounded-md before:border before:border-l-2 before:border-accent before:border-l-logo/30 before:bg-primary before:opacity-100" : "hover:bg-primary"}`}
+        className={`flex cursor-pointer select-none justify-between gap-2 px-2 py-2 ${isSelected ? "bg-primary" : "hover:bg-primary"}`}
         onClick={onClick}
         onDoubleClick={onDoubleClick}
         style={{ height: "100%" }}>
