+++ ui/src/features/Editor/components/TopBar/components/ActionBar/components/Version/index.ts
@@ -1,0 +1,1 @@
+export * from "./Version";
