+++ ui/src/lib/i18n/locales/en.json
@@ -110,15 +110,6 @@
   "Action Name": "",
   "Description": "",
   "Parameter Editor": "",
-  "Project Rollback Failed": "",
-  "Project cannot be rollbacked to this version. An error has occured.": "",
-  "Project Version Preview Creation": "",
-  "Current Version": "",
-  "Version ": "",
-  "Are you sure you want to revert to this version?": "",
-  "By clicking continue you will be reverting to version {{version}}.": "",
-  "Viewing Version": "",
-  "Version History": "",
   "Deploy Project": "",
   "Project to Deploy: ": "",
   "N/A": "",
@@ -133,9 +124,21 @@
   "Currently sharing": "",
   "Currently not sharing": "",
   "URL: ": "",
+  "Project Rollback Failed": "",
+  "Project cannot be rollbacked to this version. An error has occured.": "",
+  "Project Version Preview Creation": "",
+  "Project cannot be rolled back to this version. An error has occurred.": "",
+  "Version ": "",
+  "Are you sure you want to revert to this version?": "",
+  "By clicking continue you will be reverting to version {{version}}.": "",
+  "Viewing Version": "",
+  "Revert": "",
+  "No Version Selected": "",
+  "Current Version": "",
   "Deploy project's workflow": "",
   "Additional actions": "",
   "Export Project": "",
+  "Version History": "",
   "Stop Workflow": "",
   "Are you sure you want to stop the workflow's debug run?": "",
   "Stop": "",
