+++ ui/src/lib/i18n/locales/es.json
@@ -110,15 +110,6 @@
   "Action Name": "Nombre de la acción",
   "Description": "Descripción",
   "Parameter Editor": "Editor de parámetros",
-  "Project Rollback Failed": "",
-  "Project cannot be rollbacked to this version. An error has occured.": "",
-  "Project Version Preview Creation": "",
-  "Current Version": "Versión actual",
-  "Version ": "Versión ",
-  "Are you sure you want to revert to this version?": "¿Estás seguro de que deseas revertir a esta versión?",
-  "By clicking continue you will be reverting to version {{version}}.": "Al hacer clic en continuar, volverás a la versión {{version}}.",
-  "Viewing Version": "",
-  "Version History": "Historial de versiones",
   "Deploy Project": "Desplegar proyecto",
   "Project to Deploy: ": "Proyecto a desplegar: ",
   "N/A": "N/A",
@@ -133,9 +124,21 @@
   "Currently sharing": "Actualmente compartiendo",
   "Currently not sharing": "Actualmente no compartiendo",
   "URL: ": "URL: ",
+  "Project Rollback Failed": "",
+  "Project cannot be rollbacked to this version. An error has occured.": "",
+  "Project Version Preview Creation": "",
+  "Project cannot be rolled back to this version. An error has occurred.": "",
+  "Version ": "Versión ",
+  "Are you sure you want to revert to this version?": "¿Estás seguro de que deseas revertir a esta versión?",
+  "By clicking continue you will be reverting to version {{version}}.": "Al hacer clic en continuar, volverás a la versión {{version}}.",
+  "Viewing Version": "",
+  "Revert": "",
+  "No Version Selected": "",
+  "Current Version": "Versión actual",
   "Deploy project's workflow": "Desplegar el flujo de trabajo del proyecto",
   "Additional actions": "Acciones adicionales",
   "Export Project": "Exportar proyecto",
+  "Version History": "Historial de versiones",
   "Stop Workflow": "Detener flujo de trabajo",
   "Are you sure you want to stop the workflow's debug run?": "¿Estás seguro de que deseas detener la ejecución de depuración del flujo de trabajo?",
   "Stop": "Detener",
