+++ ui/src/lib/i18n/locales/fr.json
@@ -110,15 +110,6 @@
   "Action Name": "Nom de l'action",
   "Description": "Description",
   "Parameter Editor": "Éditeur de paramètres",
-  "Project Rollback Failed": "",
-  "Project cannot be rollbacked to this version. An error has occured.": "",
-  "Project Version Preview Creation": "",
-  "Current Version": "Version actuelle",
-  "Version ": "Version ",
-  "Are you sure you want to revert to this version?": "",
-  "By clicking continue you will be reverting to version {{version}}.": "",
-  "Viewing Version": "",
-  "Version History": "Historique des versions",
   "Deploy Project": "Déployer le projet",
   "Project to Deploy: ": "Projet à déployer: ",
   "N/A": "N/A",
@@ -133,9 +124,21 @@
   "Currently sharing": "Actuellement partagé",
   "Currently not sharing": "Actuellement non partagé",
   "URL: ": "URL: ",
+  "Project Rollback Failed": "",
+  "Project cannot be rollbacked to this version. An error has occured.": "",
+  "Project Version Preview Creation": "",
+  "Project cannot be rolled back to this version. An error has occurred.": "",
+  "Version ": "Version ",
+  "Are you sure you want to revert to this version?": "",
+  "By clicking continue you will be reverting to version {{version}}.": "",
+  "Viewing Version": "",
+  "Revert": "",
+  "No Version Selected": "",
+  "Current Version": "Version actuelle",
   "Deploy project's workflow": "",
   "Additional actions": "Actions supplémentaires",
   "Export Project": "Exporter le projet",
+  "Version History": "Historique des versions",
   "Stop Workflow": "Arrêter le flux de travail",
   "Are you sure you want to stop the workflow's debug run?": "Êtes-vous sûr de vouloir arrêter l'exécution de débogage du flux de travail?",
   "Stop": "Arrêter",
