+++ ui/src/lib/i18n/locales/ja.json
@@ -110,15 +110,6 @@
   "Action Name": "アクション名",
   "Description": "説明",
   "Parameter Editor": "パラメーターエディター",
-  "Project Rollback Failed": "プロジェクトロールバックに失敗しました",
-  "Project cannot be rollbacked to this version. An error has occured.": "プロジェクトはこのバージョンにロールバックできません。エラーが発生しました。",
-  "Project Version Preview Creation": "プロジェクトバージョンプレビュー作成",
-  "Current Version": "現在のバージョン",
-  "Version ": "バージョン ",
-  "Are you sure you want to revert to this version?": "このバージョンに戻しますか？",
-  "By clicking continue you will be reverting to version {{version}}.": "続行をクリックすると、バージョン{{version}}に戻ります。",
-  "Viewing Version": "バージョンを表示",
-  "Version History": "バージョン履歴",
   "Deploy Project": "プロジェクトをデプロイ",
   "Project to Deploy: ": "デプロイするプロジェクト: ",
   "N/A": "N/A",
@@ -133,9 +124,21 @@
   "Currently sharing": "現在共有中",
   "Currently not sharing": "現在共有していません",
   "URL: ": "URL: ",
+  "Project Rollback Failed": "プロジェクトロールバックに失敗しました",
+  "Project cannot be rollbacked to this version. An error has occured.": "プロジェクトはこのバージョンにロールバックできません。エラーが発生しました。",
+  "Project Version Preview Creation": "プロジェクトバージョンプレビュー作成",
+  "Project cannot be rolled back to this version. An error has occurred.": "",
+  "Version ": "バージョン ",
+  "Are you sure you want to revert to this version?": "このバージョンに戻しますか？",
+  "By clicking continue you will be reverting to version {{version}}.": "続行をクリックすると、バージョン{{version}}に戻ります。",
+  "Viewing Version": "バージョンを表示",
+  "Revert": "戻す",
+  "No Version Selected": "バージョンが選択されていません",
+  "Current Version": "現在のバージョン",
   "Deploy project's workflow": "プロジェクトのワークフローをデプロイ",
   "Additional actions": "追加アクション",
   "Export Project": "プロジェクトをエクスポート",
+  "Version History": "バージョン履歴",
   "Stop Workflow": "ワークフローを停止",
   "Are you sure you want to stop the workflow's debug run?": "ワークフローのデバッグ実行を停止しますか？",
   "Stop": "停止",
