+++ ui/src/lib/i18n/locales/zh.json
@@ -110,15 +110,6 @@
   "Action Name": "操作名称",
   "Description": "描述",
   "Parameter Editor": "参数编辑器",
-  "Project Rollback Failed": "",
-  "Project cannot be rollbacked to this version. An error has occured.": "",
-  "Project Version Preview Creation": "",
-  "Current Version": "当前版本",
-  "Version ": "版本 ",
-  "Are you sure you want to revert to this version?": "您确定要恢复到此版本吗？",
-  "By clicking continue you will be reverting to version {{version}}.": "点击继续，您将恢复到版本{{version}}。",
-  "Viewing Version": "",
-  "Version History": "版本历史",
   "Deploy Project": "部署项目",
   "Project to Deploy: ": "要部署的项目: ",
   "N/A": "不适用",
@@ -133,9 +124,21 @@
   "Currently sharing": "当前共享",
   "Currently not sharing": "当前未共享",
   "URL: ": "URL: ",
+  "Project Rollback Failed": "",
+  "Project cannot be rollbacked to this version. An error has occured.": "",
+  "Project Version Preview Creation": "",
+  "Project cannot be rolled back to this version. An error has occurred.": "",
+  "Version ": "版本 ",
+  "Are you sure you want to revert to this version?": "您确定要恢复到此版本吗？",
+  "By clicking continue you will be reverting to version {{version}}.": "点击继续，您将恢复到版本{{version}}。",
+  "Viewing Version": "",
+  "Revert": "",
+  "No Version Selected": "",
+  "Current Version": "当前版本",
   "Deploy project's workflow": "",
   "Additional actions": "附加操作",
   "Export Project": "导出项目",
+  "Version History": "版本历史",
   "Stop Workflow": "停止工作流",
   "Are you sure you want to stop the workflow's debug run?": "您确定要停止工作流的调试运行吗？",
   "Stop": "停止",
